+++ CHANGELOG.md
@@ -50,6 +50,7 @@
 - Fix issue with inactive locales in exports
 - Fix missing filter groups in grid filter selector when two attribute groups have the same sort orders
 - Fix mass edit issue on localizable values, it uses user locale instead of selected locale
+- Fix completeness not being correctly calculated after removing a required attribute from a family
 
 # 1.2.8 (2014-10-10)